% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd_in_tmp.R
\name{render_rmd_in_tmp}
\alias{render_rmd_in_tmp}
\title{One time setup to render R Markdown files in temporary directory}
\usage{
render_rmd_in_tmp(directory = ".")
}
\arguments{
\item{directory}{The path to the directory that contains the R Markdown files
that should be rendered in a temporary directory. Default is current
working directory.}
}
\description{
To render your R Markdown documents in a temporary directory, you need to add
a file \code{index.Rmd} that specifies \code{site: tmpsite::tmp_site} in the
YAML header. \code{build_rmd_in_tmp} creates this file.
}
\seealso{
\code{\link{tmp_site}}
}
