% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_site.R
\name{tmp_site}
\alias{tmp_site}
\title{Custom site generator to render in in a temporary directory}
\usage{
tmp_site(input, encoding = getOption("encoding"), ...)
}
\arguments{
\item{input}{Website directory (or the name of a file within the directory)}

\item{encoding}{The encoding of the input file; see \code{\link{file}}.}

\item{...}{Additional arguments (currently unused)}
}
\description{
To render your R Markdown documents in a temporary directory, you need to add a file
\code{index.Rmd} that specifies \code{site: tmpsite::tmp_site} in the YAML
header.
}
