% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_site.R
\name{tmp_site}
\alias{tmp_site}
\title{Custom site generator to render in in a temporary directory}
\usage{
tmp_site(input, encoding = getOption("encoding"), ...)
}
\arguments{
\item{input}{Website directory (or the name of a file within the directory)}

\item{encoding}{The encoding of the input file; see \code{\link{file}}.}

\item{...}{Additional arguments (currently unused)}
}
\description{
To render your R Markdown documents in a temporary directory, run
\code{\link{render_rmd_in_tmp}} in the directory that contains the R Markdown
files. Afterwards, using the RStudio Knit button will automatically use this
custom site generator. If you receive an error from RStudio, make sure the
Knit Directory is set to Document Directory (you can set this with the
dropdown menu next to the Knit button).
}
\details{
If you render the file from the R console, use
\code{\link[rmarkdown]{render_site}}. If you specify the
\code{output_format}, it is preferred to pass a character vector, e.g.
\code{"html_document"}, instead of the object itself, e.g.
\code{html_document()}.
}
\examples{
\dontrun{

library("rmarkdown")
render_site("example.Rmd", "html_document")
render_site("example.Rmd", "pdf_document")
render_site("example.Rmd", "word_document")
}
}
\seealso{
\code{\link{render_rmd_in_tmp}},
  \code{\link[rmarkdown]{render_site}}
}
